requirements.txt
openai>=1.0.0
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
numpy>=1.20.0
pydantic>=2.0.0